add payment option in product-order.php

+++ supplier-view.php
@@ -35,56 +35,56 @@ include './public/Database/update.php';
         <!-- Main Content -->
         <div class="main">
             <?php include 'partial/app-topnav.php'; ?>
-             <?php
-            $errMsg="";
+            <?php
+            $errMsg = "";
             $permissions = $_SESSION['permissions'];
             $permission_values = explode(',', $permissions[0]['permissions']);
             if (in_array("supplier-view", $permission_values)) { ?>
-            <div class="content-container">
+                <div class="content-container">
 
-                <!-- Edit Supplier Section -->
-                <div id="modal">
-                    <div class="add-user" style="background: #fff; padding: 20px; border: 1px solid black">
-                        <h3>Edit Supplier</h3>
-                        <form action="" method="POST" class="form-box">
-                            <input type="hidden" name="supplier_id" id="editId">
-                            <label>Supplier Name<input name="supplier_name" id="editName" required></label><br>
-                            <label>Supplier Location:<input name="supplier_location" id="editLocation" required></label><br>
-                            <label>Email:<input name="email" id="editEmail" type="email" required></label><br>
-                            <input class="formBtn" name="update_supplier" type="submit" value="Update">
-                            <input class="formBtn" id="cancelBtn" type="button" value="Cancel">
-                        </form>
+                    <!-- Edit Supplier Section -->
+                    <div id="modal">
+                        <div class="add-user" style="background: #fff; padding: 20px; border: 1px solid black">
+                            <h3>Edit Supplier</h3>
+                            <form action="" method="POST" class="form-box">
+                                <input type="hidden" name="supplier_id" id="editId">
+                                <label>Supplier Name<input name="supplier_name" id="editName" required></label><br>
+                                <label>Supplier Location:<input name="supplier_location" id="editLocation" required></label><br>
+                                <label>Email:<input name="email" id="editEmail" type="email" required></label><br>
+                                <input class="formBtn" name="update_supplier" type="submit" value="Update">
+                                <input class="formBtn" id="cancelBtn" type="button" value="Cancel">
+                            </form>
+                        </div>
                     </div>
-                </div>
 
-                <!-- Edit Supplier Form -->
-                <!-- List of Suppliers Section -->
-                <div class="list-users">
-                    <div class="add-user">
-                        <h3><i class="fas fa-list"></i> List of Suppliers</h3>
-                    </div>
-                    <table>
-                        <thead>
-                            <tr>
-                                <th>#</th>
-                                <th>Supplier Name</th>
-                                <th>Supplier Location</th>
-                                <th>EMAIL</th>
-                                <th>CREATED AT</th>
-                                <th>UPDATED AT</th>
-                                <th>Action</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <?php
-                            include "./public/Database/ShowData.php";
+                    <!-- Edit Supplier Form -->
+                    <!-- List of Suppliers Section -->
+                    <div class="list-users">
+                        <div class="add-user">
+                            <h3><i class="fas fa-list"></i> List of Suppliers</h3>
+                        </div>
+                        <table>
+                            <thead>
+                                <tr>
+                                    <th>#</th>
+                                    <th>Supplier Name</th>
+                                    <th>Supplier Location</th>
+                                    <th>EMAIL</th>
+                                    <th>CREATED AT</th>
+                                    <th>UPDATED AT</th>
+                                    <th>Action</th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                <?php
+                                include "./public/Database/ShowData.php";
 
-                            if (isset($_SESSION['supplier_data']) && is_array($_SESSION['supplier_data'])) {
-                                $update_data = $_SESSION['supplier_data'];
-                                $count = 0;
-                                foreach ($update_data as $supplier1) {
-                                    $supplierid = $supplier1['id'];
-                                    echo "<tr>
+                                if (isset($_SESSION['supplier_data']) && is_array($_SESSION['supplier_data'])) {
+                                    $update_data = $_SESSION['supplier_data'];
+                                    $count = 0;
+                                    foreach ($update_data as $supplier1) {
+                                        $supplierid = $supplier1['id'];
+                                        echo "<tr>
                                             <td>" . ++$count . "</td>
                                             <td>" . $supplier1['supplier_name'] . "</td>
                                             <td>" . $supplier1['supplier_location'] . "</td>
@@ -102,35 +102,36 @@ include './public/Database/update.php';
                                         <i class='fa fa-pencil'></i>Edit</a>";
                                         }
                                         if (in_array("supplier-delete", $permission_values)) {
-                                        
-                                        echo "<a href='?delete_supplier=$supplierid' class='delete-btn' onclick=\"return confirm('Are you sure you want to delete " . $supplier1["supplier_name"] . ' ' . $supplier1['supplier_location']  . "?');\"><i class='fa fa-trash'></i>Delete</a>
+
+                                            echo "<a href='?delete_supplier=$supplierid' class='delete-btn' onclick=\"return confirm('Are you sure you want to delete " . $supplier1["supplier_name"] . ' ' . $supplier1['supplier_location']  . "?');\"><i class='fa fa-trash'></i>Delete</a>
                                             </td>
                                         </tr>";
                                         }
-
+                                    }
+                                } else {
+                                    echo "<tr><td colspan='7'>No users found.</td></tr>";
                                 }
-                            } else {
-                                echo "<tr><td colspan='7'>No users found.</td></tr>";
-                            }
-                            ?>
-                        </tbody>
-                    </table>
+                                ?>
+                            </tbody>
+                        </table>
+                    </div>
                 </div>
-            </div>
-            <p class="footer-text">
+                <p class="footer-text">
+                    <?php
+                    echo "<script>console.log('Total Suppliers: " . count($update_data ?? []) . "');</script>";
+                    echo $count ?? 0;
+                    ?> Suppliers
+                </p>
                 <?php
-                echo "<script>console.log('Total Suppliers: " . count($update_data ?? []) . "');</script>";
-                echo $count ?? 0;
-                ?> Suppliers
-            </p>
-            <?php
-            if (isset($_SESSION['error'])) {
-                $error = $_SESSION['error'];
-                echo "<div id='errMsg'>$error</div>";
-                unset($_SESSION['error']);
-            }
-            ?>
-             <?php } else {
+                if (isset($_SESSION['error'])) {
+                    $error = $_SESSION['error'];
+                    if ($error != "") {
+                        echo "<div id='errMsg'>$error</div>";
+                        unset($_SESSION['error']);
+                    }
+                }
+                ?>
+            <?php } else {
                 $errMsg = "You Do not have permission to View Supplier.";
             }
